Removed the code for reading the stdout and stderr into the master daemon.
Removed the thread associated with this as well.  It was just not working there was
way too much data going through the pty's to the master and it was bogging everything
down.  We'll have to figure out another way to log messages.  Also the start up string
had to be removed because we have no way of dealing with that.  I will probably add a
local socket for master control along with some API functions for modules to communicate
with the master at some point in the future.