Fix the process restart timing code in the master daemon.
The master daemon will restart processes/modules immediately when
they die.  If the process is having a problem and will simply not
start then this would cause a high load on the system to continually
be trying to restart the process.  This timing algorithm doubles the
amount of time delay between when the process died and when it will
get restarted if the process ran for less than a minute.  If the
process runs for longer than a minute then this time will be reset
back to zero.