add code to use 'ps' as a way to retrieve cpu usage and memory usage from processes.
This is probably not the best idea in the best idea in the world
and it is not working because we have a signal handler that
waits on any child process that exits.  This causes the stream
to return EOF about half the time.  The solution may be to
implement the function with fork/pipe/exec, and record the pid
so that the cleanup function won't try to wait on the our ps
child.  There is probably still a race condition but it may be
acceptable if we deal with it gracefully.  I may forget about
it alltogether.