Add timing to the cpu and memory check in the master daemon
Don't want to run the cpu and memory check everytime we scan
the processes.  No sense it using up all those clock cycles.
The timing that we have uses the modulus of the sum of
the processes pid and a counter to have the cpu and memory
checked about every minute or so.  This won't be exact because
of signals and process restarts but that's not a big deal.  We
use the pid of the process so that we don't check the all the
processes on the same scan.  This should distribute them fairly
evenly.