Remove multiple message loop in read / write functions.
There was a race condition in the low level read/write functions that
could cause some data inconsistency if data was changed between these
multiple reads or writes.  This multiple loop function needs to be
dealt with at a higher level where the library has a better understanding
of the tag structure and can make sure that partial tags or members are
not read or written to the server.