Single socket implementation
This was to get rid of the asynchronous socket.  Before this change the
server and the clients created two sockets for communication.  One was
for synchronous messages and the other was for asynchronous events.  To
remove the async socket we had to create an event queue so that if and
event came in on the socket while we were waiting for a response to a
request we could store it for later and try again.  The event poll/wait
functions also had to change to first look in this queue and then wait
on the socket.