Removed event handling and replaced it with reading on request.
For some reason the event handling code was taking forever and causing
some errors.  I didn't really want to keep it anyway.  It would be much
more robust to read the data from OpenDAX at the moment that it was needed
instead of trying to asynchronously keep up with it using events.  At
some point I want to make it smart enough that it only reads/writes the
registers that are affected by the client request anyway and this helps
with that.