Add separate thread to handle message reception
There was a problem with using one file descriptor for message responses
and event notifications.  Before we would store events in a queue and when
the client would ask for an event it'd check the queue and if it was
empty it would then select on the fd waiting for an event.  This worked for
single threaded applications but multiple client threads would cause a
problem.  The event listener would wind up receiving the response of
normal commands and bad stuff would happen.