Add proper string handling to daxc module.  Also added comments to daxc.
If a quoted string is passed as the value in a write command it will be
handled as a string and passed to the server.  Comments were also added
and the -f argument removed in favor of simply a last argument being the
file to run. This will make it work properly with the Linux shebang #!
feature.