Change subscriptions to handle the new Lua function formatting.
Instead of trying to come up with some crazy way of dealing with data
formatting we are going to either use a straight up memory copy of the
payload or a Lua function that is defined in the configuration file
that can format the data.  At this point the function will run but
the return values aren't handled and the raw copy works.  This is the
first module to use the tag data groups.