Add fail_on_error and auto_run configurations to daxlua.  Also bug fixes
There were several bugs in daxlua, one was that the event dispatch for
the scripts was running the scripts directly instead of adding them to
the thread queue.  Once that was fixed race conditions started to show
up and those had to be fixed.  Only one of any particular script is
allowed to run at a time.