Change module tag naming philosophy.
Previously the module tag created for each module at registration was
built by adding a known unique number to the end.  This made the name
ambiguous to other modules and subject to change with the order of
module startup.  Now the number is only added to the end of the tag if
there is a collision on the name.  A warning is generated in the logs.
Should only be a problem if the same module is run more than once.  This
can be remedied by changing the modules name in the configuration.