Add special tag feature
Special tags are ones that have hooks on the read/write functions so
that the ability to be read/written in different circumstances can
be handled.  Also so that functions can be called or the data actually
manipulated in a way before it is read / written to the database.

Currently it is being used to make _module type tags read only unless
it is the tag that belongs to the calling module.