Add multi node/unit id capability to server/slave in daxmodbus
This was a fairly intrusive feature but it allows the slave or server
configurations to have multiple nodes configured so that a single port
can act like multiple devices.  The change required removing all the
tag definitions in the configuration and replacing it with the add_register
function.  If a particular register is not added to any particular node
then that will cause a wrong function code error to be returned.