Change mapping to only take one hop.  Fix bugs in mapping when writing bools.
The reason for this change was that it's not possible to have mirrored
mappings which would keep two tags equal (from a raw binary point of view)
without causing the mapping routine to recurs until the loop counter expired.
This edit causes the map to only happen once.  So one tag can be mapped
to a second but if the second tag is mapped to a third then the third
tag will not be updated if changes are made to the first.  The mapping hop
counter seems pretty clumsy and multiple maps seem to have little value.