Add user/group setting feature to daxmodbus
The reason for this change is to lower the permissions of the daxmodbus
module after the serial ports have been opened and ther server has
bound to the port.  Since the default port for Modbus is 502, root
permissions are required to open it.  Once the port is bound the user/group
can be changed to protect the system from vulnerabilities that may exist.