Various changes to tagserver
Fixed the tagserver so that it will listen to the configured
interfaces properly.

Added _tag_added and _tag_deleted tags that indicate when a tag is added
or removed from the server.

Added a feature where modules can be excluded from the creation of the
module tag via configuration options.