Change add/delete system tags to use a CDT that describes the entire tag
The _tag_added and _tag_deleted tags that get updated when a tag is added
or deleted from the server are DINTs that represent the index of the tag
that was added or deleted.  This proves to be insufficient for modules
that need to have more information.  For adding tags it's easy enough to
read the tag via the index but after a tag is deleted all the information
is lost.  Also since modules will most likely attach CHANGE events to these
tags and then immediately query the server for the tag definition, it makes
sense to include that information in these tags.